requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
manuf==1.1.5
pyfiglet==0.8.post1
python-netdiscover==0.2.0
python-nmap==0.7.1
requests==2.28.2
scapy==2.5.0
urllib3==1.26.15
